+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+Version 0.2.0 (November 17th 2020)
+-----------------------------
+ * Added support for array of different objects 
+
+Version 0.1.11 (April 10th 2020)
+-----------------------------
+ * Fixed csv export with custom labels on Firefox
+
 Version 0.1.10 (April 10th 2019)
 -----------------------------
  * Fixed csv export with better filename checking